fix error mgs using lefthook and rename setting_container.dart file
